v0.7.0 fix shop location items continuing to show up after buying them redid text logic for host text will now accept archipelago.gg:12345 as valid host removed hunie from being accumulated in the game (cant complain about it not being used if you get none) added logic for gift items to appear in the shop after receiving one made architems and rebuying unique items be in the unique items shop tab made rebuying unique items cost 0 money made giving kyu panties send location added logic for different goals (either giving kyu all panties or sleeping with all girls)

+++ Archipelago/ArchieplagoItem.cs
@@ -121,7 +121,7 @@ namespace HuniePopArchiepelagoClient.Archipelago
             return false;
         }
 
-        public bool hasitem(int flag)
+        public bool hasitem(long flag)
         {
             for (int i = 0; i < list.Count; i++)
             {
